+++ src/application/usecase/excluir_servico_usecase.py
@@ -6,4 +6,7 @@ class ExcluirServicoUsecase:
         self._servico_repository = servico_repository
 
     def executar(self, id: int):
+        servicoExiste = self._servico_repository.buscar_servico_por_id(id)
+        if not servicoExiste:
+            raise ValueError("Serviço não encontrado.")
         self._servico_repository.excluir_servico(id)
